pyproject.toml
[project]
name = "crm"
version = "0.1.0"
description = "CRM CLI for Epic Events"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[tool.poetry]
packages = [{ include = "crm", from = "src" }]